requirements.txt
attrs==19.3.0
black==19.3b0
cmake==3.16.3
efficientnet-pytorch==0.5.1
gdown==3.10.2
got10k==0.1.3
h5py==2.10.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
numpy==1.18.1
opencv-python==4.1.0.25
Pillow==7.0.0
PyYAML==5.3
scikit-image==0.16.2
scikit-learn==0.21.2
scipy==1.3.2
torch==1.4.0
torchvision==0.5.0
tqdm==4.32.2
yacs==0.1.6
youtube-dl==2020.1.24
git+https://github.com/danielgordon10/dg_util.git
